Migrate state tests to pytest
